Move executors under sglang_omni.pipeline and minor cleanups
