+++ examples/run_two_stage_demo.py
@@ -79,7 +79,7 @@ def run_stage(
     import logging
 
     from sglang_omni import Stage, Worker
-    from sglang_omni.executors import PreprocessingExecutor
+    from sglang_omni.pipeline.executor import PreprocessingExecutor
     from sglang_omni.proto import StagePayload
 
     logging.basicConfig(
