+++ examples/run_two_stage_llama_demo.py
@@ -78,7 +78,7 @@ def run_template_stage(model_id: str) -> None:
     from transformers import AutoTokenizer
 
     from sglang_omni import Stage, Worker
-    from sglang_omni.executors import PreprocessingExecutor
+    from sglang_omni.pipeline.executor import PreprocessingExecutor
     from sglang_omni.proto import StagePayload
 
     logging.basicConfig(
@@ -142,7 +142,7 @@ def run_tokenize_stage(model_id: str) -> None:
     from transformers import AutoTokenizer
 
     from sglang_omni import Stage, Worker
-    from sglang_omni.executors import PreprocessingExecutor
+    from sglang_omni.pipeline.executor import PreprocessingExecutor
     from sglang_omni.proto import StagePayload
 
     logging.basicConfig(
@@ -187,7 +187,7 @@ def run_decode_stage(model_id: str) -> None:
     from transformers import AutoTokenizer
 
     from sglang_omni import Stage, Worker
-    from sglang_omni.executors import PreprocessingExecutor
+    from sglang_omni.pipeline.executor import PreprocessingExecutor
     from sglang_omni.proto import StagePayload
 
     logging.basicConfig(
@@ -243,8 +243,8 @@ def run_engine_stage(
 
     from sglang_omni import Stage, Worker
     from sglang_omni.engines.omni import create_ar_engine
-    from sglang_omni.executors import EngineExecutor
-    from sglang_omni.executors.engine_request_builders import build_ar_request
+    from sglang_omni.pipeline.executor import EngineExecutor
+    from sglang_omni.pipeline.executor.engine_request_builders import build_ar_request
     from sglang_omni.proto import StagePayload
 
     logging.basicConfig(
