+++ sglang_omni/config/compiler.py
@@ -7,7 +7,6 @@ from pathlib import Path
 from typing import Any
 
 from sglang_omni.config.schema import InputHandlerConfig, PipelineConfig, StageConfig
-from sglang_omni.executors.interface import Executor
 from sglang_omni.pipeline import (
     AggregatedInput,
     Coordinator,
@@ -15,6 +14,7 @@ from sglang_omni.pipeline import (
     Stage,
     Worker,
 )
+from sglang_omni.pipeline.executor.interface import Executor
 from sglang_omni.pipeline.stage.input import InputHandler
 from sglang_omni.utils import import_string
