--- sglang_omni/executors/__init__.py
@@ -1,12 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-"""Executors adapt preprocessing and engines to the pipeline worker interface."""
-
-from sglang_omni.executors.engine_executor import EngineExecutor
-from sglang_omni.executors.interface import Executor
-from sglang_omni.executors.preprocessing_executor import PreprocessingExecutor
-
-__all__ = [
-    "Executor",
-    "PreprocessingExecutor",
-    "EngineExecutor",
-]
