+++ sglang_omni/pipeline/__init__.py
@@ -1,5 +1,10 @@
 # SPDX-License-Identifier: Apache-2.0
 from sglang_omni.pipeline.coordinator import Coordinator
+from sglang_omni.pipeline.executor import (
+    EngineExecutor,
+    Executor,
+    PreprocessingExecutor,
+)
 from sglang_omni.pipeline.stage.input import AggregatedInput, DirectInput, InputHandler
 from sglang_omni.pipeline.stage.runtime import Stage
 from sglang_omni.pipeline.worker.runtime import Worker
@@ -8,7 +13,12 @@ __all__ = [
     "Coordinator",
     "Stage",
     "Worker",
+    # Input handlers
     "InputHandler",
     "DirectInput",
     "AggregatedInput",
+    # Executors
+    "Executor",
+    "EngineExecutor",
+    "PreprocessingExecutor",
 ]
