+++ sglang_omni/pipeline/executor/__init__.py
@@ -1,0 +1,12 @@
+# SPDX-License-Identifier: Apache-2.0
+"""Executors adapt preprocessing and engines to the pipeline worker interface."""
+
+from sglang_omni.pipeline.executor.engine_executor import EngineExecutor
+from sglang_omni.pipeline.executor.interface import Executor
+from sglang_omni.pipeline.executor.preprocessing_executor import PreprocessingExecutor
+
+__all__ = [
+    "Executor",
+    "PreprocessingExecutor",
+    "EngineExecutor",
+]
