+++ sglang_omni/pipeline/executor/engine_request_builders.py
