+++ sglang_omni/pipeline/executor/fused_executor.py
@@ -7,7 +7,7 @@ import asyncio
 from collections.abc import Callable
 from typing import Any
 
-from sglang_omni.executors.interface import Executor
+from sglang_omni.pipeline.executor.interface import Executor
 from sglang_omni.proto import StagePayload
 from sglang_omni.utils import import_string
